Create new file
When reading the 

Regarding participation, the Directive states, "To create more informed and effective policies, the Federal Government should promote opportunities for the public to participate throughout the decision-making process. Your agency’s Open Government Plan should explain in detail how your agency will improve participation, including steps your agency will take to revise its current practices to increase opportunities for public participation in and feedback on the agency’s core mission activities." Absent from the Directive is guidance to achieve this objective; instead, the responsibility of generating civic participation is deferred to the agency. By not stipulating actual requirements, the Open Government Plan does not do anything at all to create participation. The section on collaboration reads similarly (“Your agency’s Open Government Plan should explain in detail how your agency will improve collaboration. . .”). 
In regards to transparency, the Directive requires that all government agencies create a web portal "to serve as the gateway for agency activities related to the Open Government Directive and [the agency] shall maintain and update that webpage in a timely fashion." It also has a requirement that the portal includes a mechanism for the public to provide feedback. I looked at the Department of Homeland Security's portal, and it was last updated in January 2013. Furthermore, it does not include any mechanism that allows for the public to provide feedback.    
However, in the Government's defense, participation and collaboration are the responsibility of the public. Federal policies cannot, in themselves, rectify civic engagement (or lack thereof). For instance, a 2010 study by the Pew Center states that whites are significantly more likely than blacks and latinos to participate online about government issues and policies. The multitudes of factors influencing the level of public participation are deeply rooted in history and culture, and a singular, policy-based approach will not be an effective remedy. Instead, perhaps a long-term approach, such as striving to increase technological access and literacy in specific areas of the nation, would be a more effective solution.

I agreed with many of the proposed principles outlined in the Sunlight Foundation’s Open Data Guideline, but I felt that some of the proposed principles, such as required code sharing and required documentation of data creation processes, hold the government to an unreasonably high standard. In light of Feist v. Rural Telephone Co.’s holding, it seems as if there is a lack of incentive for an organization to dedicate the time, effort, and expenses necessary for data aggregation if it the data cannot be copyrighted and other numerous users may reap the benefits of the data without contributing any of their own resources. Because the government already accepted the burden of providing data, I feel a little apprehensive hindering their data aggregation process further, since it could (in theory) stagnate the production of new data.